requirements.txt
openai
instructor
pydantic
rapidfuzz
word2number
python-dotenv